refactor the code: use the same DBsWithEnvSpecificConfig for both test cases and non-test code
